Add get_next_line and modify routes

+++ includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: aaguiler <aaguiler@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/04 18:41:38 by aaguiler          #+#    #+#             */
-/*   Updated: 2022/06/05 16:43:50 by aaguiler         ###   ########.fr       */
+/*   Updated: 2022/06/07 16:23:12 by aaguiler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ typedef struct s_stacks
 int			ft_atoi(char *str, t_stacks *stacks);
 char		**ft_split(char const *s, char c);
 int			ft_count_words(char const *s, char c);
-void		*ft_calloc(size_t count, size_t size);
+void		*ft_calloc(int count, int size);
 
 //Operations functions
 void		ft_swap_a(t_stacks *stacks, int print);
